+++ requestly-android-core/src/main/java/io/requestly/android/core/modules/apiModifier/processors/models/Action.kt
@@ -1,4 +1,4 @@
-package io.requestly.android.core.modules.apiModifier.models
+package io.requestly.android.core.modules.apiModifier.processors.models
 
 import java.net.URL
